Added lawyer register/auth & fixed homepage
