+++ frontend/pages/cases.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function cases() {
+    return <div>Cases</div>;
+}
+
+export default cases;
