+++ frontend/pages/property/index.jsx
@@ -1,13 +1,13 @@
-import Header from "@/components/Header"
-import Footer from "@/components/Footer"
-import GetStarted from "@/components/property/getStarted"
+import Header from "@/components/Header";
+import Footer from "@/components/Footer";
+import GetStarted from "@/components/property/getStarted";
 
 export default function Property() {
     return (
-      <div>
-        <Header />
-      <GetStarted />
-        <Footer />
-      </div>
-    )
-  }
+        <div>
+            <Header />
+            <GetStarted />
+            <Footer />
+        </div>
+    );
+}
